Moved ep40 to ep/40 for easier adding of next ep keyboard (#4439)
* add initial ep40 files

* fixed issues

* updated keymap

* Added media control

* Update keyboards/handwired/ep40/rules.mk


* Fixed requested changes

* Fixed more requested changes

* Added delete key to layor 1

* Updated defualt keympap to have a Backspace mod del key

* removed place holder


* removed obsolete code


* Moved ep40 to ep/40

* Fixed file names for moved code

* Fixed readme
