+++ keyboards/ep/40/40.c
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "ep40.h"
+#include "40.h"
 
 void matrix_init_kb(void) {
 	// put your keyboard start-up code here
