+++ keyboards/ep/40/readme.md
@@ -4,12 +4,12 @@
 
 A simple 40% keyboard
 
-Keyboard Maintainer: [Elliot Powell](https://reddit.com/u/e11i0t23, https://github.com/e11i0t23)  
+Keyboard Maintainer: [Elliot Powell](https://github.com/e11i0t23), [/u/e11i0t23 on reddit](https://reddit.com/u/e11i0t23)  
 Hardware Supported: EP40 PCB  
 Hardware Availability: None at the moment
 
 Make example for this keyboard (after setting up your build environment):
 
-    make ep40:default
+    make ep/40:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
